+++ system/lib/standalone/standalone.c
@@ -74,14 +74,14 @@ weak int _mmap_js(size_t length,
                   int prot,
                   int flags,
                   int fd,
-                  size_t offset,
+                  off_t offset,
                   int* allocated,
                   void** addr) {
   return -ENOSYS;
 }
 
 weak int _munmap_js(
-  intptr_t addr, size_t length, int prot, int flags, int fd, size_t offset) {
+  intptr_t addr, size_t length, int prot, int flags, int fd, off_t offset) {
   return -ENOSYS;
 }
